feat: advisory lock on core partitions
