+++ pkg/repository/sqlchelpers/tx.go
@@ -93,51 +93,15 @@ func PrepareTxWithStatementTimeout(ctx context.Context, pool *pgxpool.Pool, l *z
 	return tx, commit, rollback, nil
 }
 
-// AcquireConnectionWithStatementTimeout acquires a connection from the pool and overwrites the default statement timeout on it.
-// It does not support timeout values lower than the default timeout (if called with such a value, it will just use the default timeout).
+// AcquireConnectionWithStatementTimeout acquires a connection from the pool and overwrites the statement timeout on it.
 func AcquireConnectionWithStatementTimeout(ctx context.Context, pool *pgxpool.Pool, l *zerolog.Logger, timeoutMs int) (*pgx.Conn, func(), error) {
-	start := time.Now()
-
-	conn, err := pool.Acquire(ctx)
-
+	conn, err := AcquirePoolConnectionWithStatementTimeout(ctx, pool, l, timeoutMs)
 	if err != nil {
-		if sinceStart := time.Since(start); sinceStart > 100*time.Millisecond {
-			l.Error().Dur(
-				"duration", sinceStart,
-			).Int(
-				"acquired_connections", int(pool.Stat().AcquiredConns()),
-			).Caller(1).Msgf("long connection acquire with error: %v", err)
-		}
-
 		return nil, nil, err
-	}
-
-	if sinceStart := time.Since(start); sinceStart > 100*time.Millisecond {
-		l.Warn().Dur(
-			"duration", sinceStart,
-		).Int(
-			"acquired_connections", int(pool.Stat().AcquiredConns()),
-		).Caller(1).Msg("long connection acquire")
 	}
 
 	release := func() {
-		// reset statement timeout with a separate ctx; we don't want to use the original ctx here in case it has been cancelled
-		resetCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, err = conn.Exec(resetCtx, "SET statement_timeout=30000")
-
-		if err != nil {
-			l.Error().Err(err).Msg("failed to reset statement timeout on released connection")
-		}
-
-		conn.Release()
-	}
-
-	_, err = conn.Exec(ctx, fmt.Sprintf("SET statement_timeout=%d", timeoutMs))
-
-	if err != nil {
-		release()
-		return nil, nil, err
+		ReleasePoolConnectionWithStatementTimeout(conn, l)
 	}
 
 	return conn.Conn(), release, nil
